Create IMU and UWB data(Code)
The given dataset is sensor measurements IMU and UWB data. Each line represents an individual sample and columns to the three axes of X, Y, Z components of a sensor in IMU along with UWB data. The IMU data specifies the acceleration or orientation in every axis, while it seems more likely that UWB calculates distance and position. Each cell value represents the observed quantity of each sample. The data collection was likely done by putting the sensors on a device or subject and reading them when moving when resting. This data set is essential for understanding the dynamics around IMU and UWB technologies; it offers crucial details on spatial parameters along with motion.